+++ src/xslint.js
@@ -14,6 +14,7 @@ const {lintByXpath, names: xpathChecks} = require('./xpath-linter')
 const {lintByCorpus, names: corpusChecks} = require('./corpus-linter')
 const {lintByAxis, names: axisChecks} = require('./xpath-axis-linter')
 const {lintByNamespace, names: namespaceChecks} = require('./namespace-linter')
+const {lintByNodeSet, names: nodeSetChecks} = require('./node-set-linter')
 const {lintByFormat, names: formatChecks} = require('./xpath-format-linter')
 const {fixed} = require('./fixer')
 const {logger, levels} = require('./logger')
@@ -32,6 +33,7 @@ const LINTERS = [
   lintByCorpus,
   lintByAxis,
   lintByNamespace,
+  lintByNodeSet,
 ]
 
 /**
@@ -52,7 +54,7 @@ const EXPRESSION_LINTERS = [
 const CHECKS = [
   ...xslChecks, ...xpathValidatorChecks,
   ...xpathChecks, ...corpusChecks, ...axisChecks, ...namespaceChecks,
-  ...formatChecks,
+  ...nodeSetChecks, ...formatChecks,
 ]
 
 /**