+++ test/node-set-linter.test.js
@@ -1,0 +1,34 @@
+/*
+ * SPDX-FileCopyrightText: Copyright (c) 2025-2026 Max Trunnikov
+ * SPDX-License-Identifier: MIT
+ */
+
+const {lintByNodeSet} = require('../src/node-set-linter')
+const {allFilesFrom, xml, yaml} = require('../src/helpers')
+const path = require('path')
+const assert = require('assert')
+
+/**
+ * Yaml node-set linter test packs.
+ * @type {Array<string>}
+ */
+const PACKS = allFilesFrom(
+  path.resolve(__dirname, 'resources', 'node-set-packs'),
+)
+
+describe('node-set-linter', function() {
+  PACKS.forEach((pack) => {
+    const yml = yaml.parsedFromFile(pack)
+    const input = xml.parsedFromString(yml.input)
+    describe(`testing ${path.basename(pack)} pack`, function() {
+      it(`should find ${yml.found.amount} node-set extensions`, function() {
+        const defects = lintByNodeSet([{file: 'test.xsl', xsl: input}])
+        assert.equal(defects.length, yml.found.amount)
+        yml.found.positions.forEach((pos, index) => {
+          assert.equal(defects[index].line, pos[0])
+          assert.equal(defects[index].pos, pos[1])
+        })
+      })
+    })
+  })
+})